requirements.txt
facenet_pytorch==2.5.3
Flask==3.0.2
numpy==1.26.4
opencv_python==4.9.0.80
python-dotenv==1.0.1
torch==2.2.1
ultralytics==8.1.27
